Fix: Delete files of mypage in auth-server
Account, User model과 AccountSerializer, UserCreateSerializer
제외하고 모두 삭제
