+++ auth-server/config/urls.py
@@ -36,8 +36,6 @@ urlpatterns = [
     path('user/', include([
         # path('admin/', admin.site.urls),
         path('', include('accounts.urls')),
-        path('', include('mypage.urls')),
-        # path(r'swagger(?P<format>\.json|\.yaml)', schema_view.without_ui(cache_timeout=0), name='schema-json'),
         path('swagger.yaml', schema_view.without_ui(cache_timeout=0), name='schema-yaml'),
         path('swagger', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
     ]))
