+++ auth-server/mypage/models.py
@@ -39,90 +39,3 @@ class Account(models.Model):
     class Meta:
         managed = TEST
         db_table = 'Account'
-
-
-class Bookmark(models.Model):
-    user = models.ForeignKey('User', models.DO_NOTHING, related_name='bookmarks')
-    ticket = models.ForeignKey('Ticket', models.DO_NOTHING, related_name='bookmarks')
-
-    class Meta:
-        managed = TEST
-        db_table = 'Bookmark'
-        unique_together = (('user', 'ticket'),)
-
-
-class Buy(models.Model):
-    user = models.ForeignKey('User', models.DO_NOTHING, related_name='buys')
-    ticket = models.OneToOneField('Ticket', on_delete=models.DO_NOTHING)
-    date = models.DateTimeField(blank=True, null=True, help_text='구매일시')
-
-    class Meta:
-        managed = TEST
-        db_table = 'Buy'
-        ordering = ['-date']
-
-
-class Ticket(models.Model):
-    seller = models.ForeignKey('User', models.DO_NOTHING, related_name='sell_tickets')
-    location = models.CharField(max_length=255)
-    price = models.IntegerField()
-    created_at = models.DateTimeField()
-    state = models.IntegerField(help_text='양도권 상태 (0: 판매중, 1: 예약중, 2: 판매완료)')
-    tag_hash = models.BigIntegerField()
-    is_membership = models.BooleanField()
-    is_holding = models.BooleanField()
-    remaining_number = models.IntegerField(blank=True, null=True)
-    type = models.IntegerField()
-    can_nego = models.BooleanField()
-    trade_type = models.IntegerField()
-    has_shower = models.BooleanField()
-    has_locker = models.BooleanField()
-    has_clothes = models.BooleanField()
-    has_gx = models.BooleanField()
-    can_resell = models.BooleanField()
-    can_refund = models.BooleanField()
-    description = models.CharField(max_length=255)
-    thumbnail = models.CharField(max_length=255, blank=True, null=True)
-    transfer_fee = models.IntegerField()
-    # point = models.TextField()  # This field type is a guess.
-    address = models.CharField(max_length=255)
-    main_image = models.CharField(max_length=255, blank=True, null=True)
-    expiry_date = models.DateField(blank=True, null=True)
-
-    buyer = models.ForeignKey('User', models.DO_NOTHING, blank=True, null=True, related_name='buy_tickets')
-    bookmark_users = models.ManyToManyField('User', through='Bookmark', related_name='bookmark_tickets')
-    tags = models.ManyToManyField('Tag', through='TicketTag', related_name='tickets')
-
-    class Meta:
-        managed = TEST
-        db_table = 'Ticket'
-        ordering = ['-created_at']
-
-
-class TicketImage(models.Model):
-    ticket = models.ForeignKey(Ticket, models.DO_NOTHING, related_name='images')
-    url = models.CharField(max_length=255)
-    is_main = models.BooleanField()
-    thumbnail_url = models.CharField(max_length=255)
-
-    class Meta:
-        managed = TEST
-        db_table = 'TicketImage'
-
-
-class Tag(models.Model):
-    subject = models.CharField(max_length=255)
-    content = models.TextField()
-
-    class Meta:
-        managed = TEST
-        db_table = 'Tag'
-
-
-class TicketTag(models.Model):
-    ticket = models.ForeignKey('Ticket', models.DO_NOTHING)
-    tag = models.ForeignKey('Tag', models.DO_NOTHING)
-
-    class Meta:
-        managed = TEST
-        db_table = 'TicketTag'
